+++ tests/test_dependencies.py
@@ -18,7 +18,7 @@ import os
 
 G_LOGGER.verbosity = logger.Verbosity.VERBOSE
 
-SBUILDR_ROOT = os.path.abspath(os.path.join(TESTS_ROOT, os.pardir))
+SBUILDR_ROOT = os.path.abspath(os.path.join(TESTS_ROOT, os.path.pardir))
 sys.path.insert(0, SBUILDR_ROOT)
 
 class TestGitFetcher(object):